+++ src/clojisr/v1/codegen.clj
@@ -23,7 +23,7 @@
   (map #(some-> % f) xs))
 
 ;; all binary operators as set of strings
-(def binary-operators #{"$" "=" "<<-" "<-" "+" "-" "/" "*" "&" "&&" "|" "||" "==" "!=" "<=" ">=" "<" ">"})
+(def binary-operators #{"$" "=" "<<-" "<-" "+" "-" "/" "*" "&" "&&" "|" "||" "==" "!=" "<=" ">=" "<" ">" "%in%" "%%" "%/%" "**"})
 (def binary-operators-flat #{"=" "<<-" "<-"})
 (def unary-operators #{"+" "-"})
 (def wrapped-operators #{"+" "-" "/" "*" "&" "&&" "|" "||" "==" "!=" "<=" ">=" "<" ">"})